Added audio and options menu
sound effects for zombie attacking, building destruction, allies (soldier and turrets) attacking and night starting.
music for main menu and in game.
options to lower and raise volume of both sound effects and music as well as to each one individually.
Downloaded particles for shooting but not added it yet to turrets and soldiers.